+++ src/test/java/com/example/solidconnection/e2e/ApplicantsQueryTest.java
@@ -7,8 +7,7 @@ import com.example.solidconnection.application.dto.ApplicantResponse;
 import com.example.solidconnection.application.dto.ApplicationsResponse;
 import com.example.solidconnection.application.dto.UniversityApplicantsResponse;
 import com.example.solidconnection.application.repository.ApplicationRepository;
-import com.example.solidconnection.config.token.TokenProvider;
-import com.example.solidconnection.config.token.TokenType;
+import com.example.solidconnection.auth.service.TokenProvider;
 import com.example.solidconnection.auth.domain.TokenType;
 import com.example.solidconnection.siteuser.domain.SiteUser;
 import com.example.solidconnection.siteuser.repository.SiteUserRepository;
